[Rector] Apply Rector dead code set list

+++ system/Database/MySQLi/Result.php
@@ -89,7 +89,7 @@ class Result extends BaseResult
             $retVal[$i]->type        = $data->type;
             $retVal[$i]->type_name   = in_array($data->type, [1, 247], true) ? 'char' : ($dataTypes[$data->type] ?? null);
             $retVal[$i]->max_length  = $data->max_length;
-            $retVal[$i]->primary_key = (int) ($data->flags & 2);
+            $retVal[$i]->primary_key = $data->flags & 2;
             $retVal[$i]->length      = $data->length;
             $retVal[$i]->default     = $data->def;
         }
