+++ system/Debug/Toolbar/Collectors/Files.php
@@ -45,7 +45,7 @@ class Files extends BaseCollector
      */
     public function getTitleDetails(): string
     {
-        return '( ' . (int) count(get_included_files()) . ' )';
+        return '( ' . count(get_included_files()) . ' )';
     }
 
     /**
